Unversion opt/mattermost
The idea behind opt/mattermost-<ver> was to "version" configuration and
data so that it would be easier-ish to move between versions and
providing an easier upgrade path. This has turned out to be more work
than it is beneficial, as the important configuration is git tracked,
and bumping version numbers mean updating more files than necessary.

The auxillary files in opt/ still needs to be backed up and potentially
rolled back **if** an upgrade would happen to wreck havoc on the
directory, but this is probably better achieved the other way aroud,
that is, snapshots are made when upgrading and the update still uses the
same dir.