Add mattermost guix package, entry point
Add a package definition for guix [1] to help manage the mattermost
version, as it got tedious already after three upgrades. The packaging
itself is pretty trivial as the pre-built mattermost binaries already
bundle everything they need, and the "build" is simply a copy.

The most notable effect is the work when upgrading. With guix, you need
to bump the version in opt/mattermost/mattermost.scm, update the hash,
and push the changes to the server (with rsync, for now). Then,
restarting the systemd service should pull the new version, although to
keep downtime to a minimum it might be a good idea to pre-build the
package.

The package only maintains the mattermost binaries - the data/, logs/
etc. are mostly separate like before.

[1] https://guix.gnu.org/