Version control the authorized_keys for root
Add `root/.ssh/authorized_keys` file with @jokva and @santisoler ssh
keys, and install them to the root dir with stow.

This change makes it so that the default authorized_keys, the one input
in the hetzner console when creating the instance, will *lose access*
unless they are also added to the authorized_keys in the tree *before*
running init. While a sharp edge, at that point re-initializing is
cheap, and is likely to be done by a maintainer anyway.