Revert "Don't escape & in postgres connect string"
This reverts commit e8da8a325303caa3b8c42212e9d174272737c243.

Going over the logs again, the problem was never really the \u0026, but
rather a bad config being used by an (experimental, now discarded)
startup script, and the whole connection string was wrong. I pulled the
config file after restarting Mattermost 9.11.7 and it had replaced all
& with \u0026 in the config file. It therefore makes sense to revert
this commit so that we can compare generations of the config file
without this noise.